requirements.txt
langchain>=0.3.25
langgraph>=0.2.39
langchain-openai>=0.2.14
langchain-ollama>=0.2.2
python-dotenv>=1.0.1
